core: Remove CORS logic
