+++ examples/examples/cors_server.rs
@@ -68,7 +68,7 @@ async fn main() -> anyhow::Result<()> {
 }
 
 async fn run_server() -> anyhow::Result<(SocketAddr, HttpServerHandle)> {
-	let acl = AccessControlBuilder::new().allow_all_headers().allow_all_origins().allow_all_hosts().build();
+	let acl = AccessControlBuilder::new().allow_all_origins().allow_all_hosts().build();
 
 	let server =
 		HttpServerBuilder::default().set_access_control(acl).build("127.0.0.1:0".parse::<SocketAddr>()?).await?;
